Added new functionalities FE-BE

+++ Student-Service-BE/src/main/java/student/service/service/impl/CityServiceImpl.java
@@ -4,22 +4,17 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import javax.persistence.EntityExistsException;
+import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 
-import org.mapstruct.factory.Mappers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import student.service.dao.CityDao;
-import student.service.dto.CityDto;
-import student.service.dto.StudentDto;
 import student.service.entity.CityEntity;
-import student.service.entity.StudentEntity;
-import student.service.exception.EntityNotPresent;
-import student.service.exception.ExistEntityException;
-import student.service.mapper.CityMapper;
 import student.service.service.CityService;
 
 @Service
@@ -28,8 +23,6 @@ public class CityServiceImpl implements CityService {
 
 	private final CityDao cityDao;
 
-	private final CityMapper cityMapper = Mappers.getMapper(CityMapper.class);
-
 	@Autowired
 	public CityServiceImpl(CityDao cityDao) {
 		super();
@@ -37,50 +30,53 @@ public class CityServiceImpl implements CityService {
 	}
 
 	@Override
-	public List<CityDto> findAll() {
-		List<CityEntity> entites = (List)cityDao.findAll();
+	public List<CityEntity> findAll() {
+		List<CityEntity> entites = (List<CityEntity>)cityDao.findAll();
 		return entites.stream().map(entity -> {
-			return cityMapper.toCityDto(entity);
+			return entity;
 		}).collect(Collectors.toList());
 	}
 	
 	@Override
-	public Page<CityDto> findByPage(Pageable pageable) {
-		Page<CityDto> entites = cityDao.findAll(pageable).map(cityMapper::toCityDto);
+	public Page<CityEntity> findByPage(Pageable pageable) {
+		Page<CityEntity> entites = cityDao.findAll(pageable);
 		return entites;
 	}
 
 	@Override
-	public void deleteById(Long id) {
-		cityDao.deleteById(id);
+	public void deleteById(Long id) throws EntityNotFoundException {
+		if(cityDao.existsById(id))
+			cityDao.deleteById(id);
+		else
+			throw new EntityNotFoundException("City with that ID does not exists!");
 	}
 
 	@Override
-	public CityDto save(CityDto cityDto) throws ExistEntityException {
-		Optional<CityEntity> entity = cityDao.findById(cityDto.getNumber());
+	public CityEntity save(CityEntity cityEntity) throws EntityExistsException {
+		Optional<CityEntity> entity = cityDao.findById(cityEntity.getNumber());
 		if (entity.isPresent()) {
-			throw new ExistEntityException(entity.get(), "City already exists!");
+			throw new EntityExistsException("City already exists!");
 		}
-		CityEntity c = cityDao.save(cityMapper.toCity(cityDto));
-		return cityMapper.toCityDto(c);
+		CityEntity c = cityDao.save(cityEntity);
+		return c;
 	}
 	
 	@Override
-	public CityDto update(CityDto cityDto) throws EntityNotPresent {
-		Optional<CityEntity> entity = cityDao.findById(cityDto.getNumber());
+	public CityEntity update(CityEntity cityEntity) throws EntityNotFoundException {
+		Optional<CityEntity> entity = cityDao.findById(cityEntity.getNumber());
 		if (!entity.isPresent()) {
-			throw new EntityNotPresent(cityDto.getNumber(), "City does not exist!");
+			throw new EntityNotFoundException("City does not exist!");
 		}
-		CityEntity c = cityDao.save(cityMapper.toCity(cityDto));
-		return cityMapper.toCityDto(c);
+		CityEntity c = cityDao.save(cityEntity);
+		return c;
 		
 	}
 
 	@Override
-	public Optional<CityDto> findById(Long id) {
+	public Optional<CityEntity> findById(Long id) {
 		Optional<CityEntity> city = cityDao.findById(id);
 		if (city.isPresent()) {
-			return Optional.of(cityMapper.toCityDto(city.get()));
+			return Optional.of(city.get());
 		}
 		return Optional.empty();
 	}
