+++ Student-Service-BE/src/main/java/student/service/service/impl/ExamRegistrationServiceImpl.java
@@ -1,23 +1,21 @@
 package student.service.service.impl;
 
+import java.sql.Date;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import javax.persistence.EntityExistsException;
+import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 
-import org.mapstruct.factory.Mappers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import student.service.dao.ExamRegistrationDao;
-import student.service.dto.ExamRegistrationDto;
 import student.service.entity.ExamRegistrationEntity;
-import student.service.exception.EntityNotPresent;
-import student.service.exception.ExistEntityException;
-import student.service.mapper.ExamRegistrationMapper;
 import student.service.service.ExamRegistrationService;
 
 @Service
@@ -26,60 +24,66 @@ public class ExamRegistrationServiceImpl implements ExamRegistrationService {
 
 	private final ExamRegistrationDao examRegistrationDao;
 
-	private final ExamRegistrationMapper examRegistrationMapper = Mappers.getMapper(ExamRegistrationMapper.class);
-
 	@Autowired
 	public ExamRegistrationServiceImpl(ExamRegistrationDao examRegistrationDao) {
 		super();
 		this.examRegistrationDao = examRegistrationDao;
 	}
 
-	
 	@Override
-	public List<ExamRegistrationDto> findAll() {
-		List<ExamRegistrationEntity> entites = (List)examRegistrationDao.findAll();
+	public List<ExamRegistrationEntity> findAll() {
+		List<ExamRegistrationEntity> entites = (List<ExamRegistrationEntity>)examRegistrationDao.findAll();
 		return entites.stream().map(entity -> {
-			return examRegistrationMapper.toExamRegistrationDto(entity);
+			return (entity);
 		}).collect(Collectors.toList());
 	}
 	
 	@Override
-	public Page<ExamRegistrationDto> findByPage(Pageable pageable) {
-		Page<ExamRegistrationDto> entites = examRegistrationDao.findAll(pageable).map(examRegistrationMapper::toExamRegistrationDto);
-		return entites;
+	public Page<ExamRegistrationEntity> findByPage(Pageable pageable, String search) {
+		if (search == null)
+			return examRegistrationDao.findAll(pageable);
+		else
+			return examRegistrationDao.findByStudentIndexNumberContaining(pageable, search);
 	}
 
 	@Override
-	public void deleteById(Long id) {
-		examRegistrationDao.deleteById(id);
+	public void deleteById(Long id) throws EntityNotFoundException {
+		if(examRegistrationDao.existsById(id))
+			examRegistrationDao.deleteById(id);
+		else
+			throw new EntityNotFoundException("Exam registration with that ID does not exists!");
 	}
 
 	@Override
-	public ExamRegistrationDto save(ExamRegistrationDto examDto) throws ExistEntityException {
-		Optional<ExamRegistrationEntity> entity = examRegistrationDao.findById(examDto.getId());
+	public ExamRegistrationEntity save(ExamRegistrationEntity examRegistrationEntity) throws EntityExistsException {
+		Optional<ExamRegistrationEntity> entity = examRegistrationDao.findByStudentAndExam(examRegistrationEntity.getStudent().getIndexNumber(), examRegistrationEntity.getExam().getId());
 		if (entity.isPresent()) {
-			throw new ExistEntityException(entity.get(), "Exam registration already exists!");
+			throw new EntityExistsException("You already registered for that exam!");
 		}
-		ExamRegistrationEntity e = examRegistrationDao.save(examRegistrationMapper.toExamRegistration(examDto));
-		return examRegistrationMapper.toExamRegistrationDto(e);
+		
+		long millis=System.currentTimeMillis();  
+        Date date = new Date(millis);      
+		examRegistrationEntity.setExamRegistrationTime(date);
+		ExamRegistrationEntity e = examRegistrationDao.save(examRegistrationEntity);
+		return e;
 	}
 	
 	@Override
-	public ExamRegistrationDto update(ExamRegistrationDto examDto) throws EntityNotPresent {
-		Optional<ExamRegistrationEntity> entity = examRegistrationDao.findById(examDto.getId());
+	public ExamRegistrationEntity update(ExamRegistrationEntity examRegistrationEntity) throws EntityNotFoundException {
+		Optional<ExamRegistrationEntity> entity = examRegistrationDao.findById(examRegistrationEntity.getId());
 		if (!entity.isPresent()) {
-			throw new EntityNotPresent(examDto.getId(), "Exam registration does not exist!");
+			throw new EntityNotFoundException("Exam registration does not exist!");
 		}
-		ExamRegistrationEntity e = examRegistrationDao.save(examRegistrationMapper.toExamRegistration(examDto));
-		return examRegistrationMapper.toExamRegistrationDto(e);
+		ExamRegistrationEntity e = examRegistrationDao.save(examRegistrationEntity);
+		return e;
 		
 	}
 	
 	@Override
-	public Optional<ExamRegistrationDto> findById(Long id) {
-		Optional<ExamRegistrationEntity> exam = examRegistrationDao.findById(id);
-		if (exam.isPresent()) {
-			return Optional.of(examRegistrationMapper.toExamRegistrationDto(exam.get()));
+	public Optional<ExamRegistrationEntity> findById(Long id) {
+		Optional<ExamRegistrationEntity> examRegistrationEntity = examRegistrationDao.findById(id);
+		if (examRegistrationEntity.isPresent()) {
+			return Optional.of(examRegistrationEntity.get());
 		}
 		return Optional.empty();
 	}
