+++ Student-Service-BE/src/main/java/student/service/service/impl/ExamServiceImpl.java
@@ -4,21 +4,17 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import javax.persistence.EntityExistsException;
+import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 
-import org.mapstruct.factory.Mappers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import student.service.dao.ExamDao;
-import student.service.dto.ExamDto;
 import student.service.entity.ExamEntity;
-import student.service.exception.EntityNotPresent;
-import student.service.exception.ExistEntityException;
-import student.service.mapper.ExamMapper;
-import student.service.mapper.ProfessorMapper;
 import student.service.service.ExamService;
 
 @Service
@@ -27,60 +23,71 @@ public class ExamServiceImpl implements ExamService {
 
 	private final ExamDao examDao;
 
-	private final ExamMapper examMapper = Mappers.getMapper(ExamMapper.class);
-
 	@Autowired
 	public ExamServiceImpl(ExamDao examDao) {
 		super();
 		this.examDao = examDao;
 	}
-
 	
 	@Override
-	public List<ExamDto> findAll() {
-		List<ExamEntity> entites = (List)examDao.findAll();
+	public List<ExamEntity> findAll() {
+		List<ExamEntity> entites = (List<ExamEntity>)examDao.findAll();
 		return entites.stream().map(entity -> {
-			return examMapper.toExamDto(entity);
+			return (entity);
 		}).collect(Collectors.toList());
 	}
 	
 	@Override
-	public Page<ExamDto> findByPage(Pageable pageable) {
-		Page<ExamDto> entites = examDao.findAll(pageable).map(examMapper::toExamDto);
-		return entites;
+	public Page<ExamEntity> findByPage(Pageable pageable, String search) {
+		if (search == null)
+			return examDao.findAll(pageable);
+		else
+			return examDao.findBySubjectNameContaining(pageable, search);
 	}
 
 	@Override
-	public void deleteById(Long id) {
-		examDao.deleteById(id);
+	public void deleteById(Long id) throws EntityNotFoundException {
+		if(examDao.existsById(id))
+			examDao.deleteById(id);
+		else
+			throw new EntityNotFoundException("Exam with that ID does not exists!");
 	}
 
 	@Override
-	public ExamDto save(ExamDto examDto) throws ExistEntityException {
-		Optional<ExamEntity> entity = examDao.findById(examDto.getSubjectDto().getId());
+	public ExamEntity save(ExamEntity examEntity) throws EntityExistsException {
+		Optional<ExamEntity> entity = examDao.findExamWithSubjectAndDate(examEntity.getSubject().getId());
 		if (entity.isPresent()) {
-			throw new ExistEntityException(entity.get(), "Exam already exists!");
+			throw new EntityExistsException("There is currently an active exam for the selected subject!");
 		}
-		ExamEntity e = examDao.save(examMapper.toExam(examDto));
-		return examMapper.toExamDto(e);
+		ExamEntity e = examDao.save(examEntity);
+		return e;
 	}
 	
 	@Override
-	public ExamDto update(ExamDto examDto) throws EntityNotPresent {
-		Optional<ExamEntity> entity = examDao.findById(examDto.getId());
+	public ExamEntity update(ExamEntity examEntity) throws EntityNotFoundException {
+		Optional<ExamEntity> entity = examDao.findById(examEntity.getId());
 		if (!entity.isPresent()) {
-			throw new EntityNotPresent(examDto.getId(), "Exam does not exist!");
+			throw new EntityNotFoundException("Exam does not exist!");
 		}
-		ExamEntity e = examDao.save(examMapper.toExam(examDto));
-		return examMapper.toExamDto(e);
+		ExamEntity e = examDao.save(examEntity);
+		return e;
 		
 	}
 	
 	@Override
-	public Optional<ExamDto> findById(Long id) {
+	public Optional<ExamEntity> findById(Long id) {
 		Optional<ExamEntity> exam = examDao.findById(id);
 		if (exam.isPresent()) {
-			return Optional.of(examMapper.toExamDto(exam.get()));
+			return Optional.of(exam.get());
+		}
+		return Optional.empty();
+	}
+
+	@Override
+	public Optional<List<?>> findAllPosibleStudentExams(String indexNumber) {
+		Optional<List<?>> exam = examDao.findAllPosibleStudentExams(indexNumber);
+		if (exam.isPresent()) {
+			return Optional.of(exam.get());
 		}
 		return Optional.empty();
 	}
