+++ Student-Service-BE/src/main/java/student/service/service/impl/TitleServiceImpl.java
@@ -4,21 +4,17 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import javax.persistence.EntityExistsException;
+import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 
-import org.mapstruct.factory.Mappers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import student.service.dao.TitleDao;
-import student.service.dto.CityDto;
-import student.service.dto.TitleDto;
 import student.service.entity.TitleEntity;
-import student.service.exception.EntityNotPresent;
-import student.service.exception.ExistEntityException;
-import student.service.mapper.TitleMapper;
 import student.service.service.TitleService;
 
 @Service
@@ -27,8 +23,6 @@ public class TitleServiceImpl implements TitleService {
 
 	private final TitleDao titleDao;
 
-	private final TitleMapper titleMapper = Mappers.getMapper(TitleMapper.class);
-
 	@Autowired
 	public TitleServiceImpl(TitleDao titleDao) {
 		super();
@@ -36,50 +30,52 @@ public class TitleServiceImpl implements TitleService {
 	}
 
 	@Override
-	public List<TitleDto> findAll() {
-		List<TitleEntity> entites = (List)titleDao.findAll();
+	public List<TitleEntity> findAll() {
+		List<TitleEntity> entites = (List<TitleEntity>)titleDao.findAll();
 		return entites.stream().map(entity -> {
-			return titleMapper.toTitleDto(entity);
+			return (entity);
 		}).collect(Collectors.toList());
 	}
 	
 	@Override
-	public Page<TitleDto> findByPage(Pageable pageable) {
-		Page<TitleDto> entites = titleDao.findAll(pageable).map(titleMapper::toTitleDto);
+	public Page<TitleEntity> findByPage(Pageable pageable) {
+		Page<TitleEntity> entites = titleDao.findAll(pageable);
 		return entites;
 	}
 
 	@Override
-	public void deleteById(Long id) {
-		titleDao.deleteById(id);
+	public void deleteById(Long id) throws EntityNotFoundException {
+		if(titleDao.existsById(id))
+			titleDao.deleteById(id);
+		else
+			throw new EntityNotFoundException("Title with that ID does not exists!");
 	}
 
 	@Override
-	public TitleDto save(TitleDto titleDto) throws ExistEntityException {
-		Optional<TitleEntity> entity = titleDao.findById(titleDto.getId());
+	public TitleEntity save(TitleEntity titleEntity) throws EntityExistsException {
+		Optional<TitleEntity> entity = titleDao.findById(titleEntity.getId());
 		if (entity.isPresent()) {
-			throw new ExistEntityException(entity.get(), "Title already exists!");
+			throw new EntityExistsException("Title already exists!");
 		}
-		TitleEntity t = titleDao.save(titleMapper.toTitle(titleDto));
-		return titleMapper.toTitleDto(t);
+		TitleEntity t = titleDao.save(titleEntity);
+		return t;
 	}
 	
 	@Override
-	public TitleDto update(TitleDto titleDto) throws EntityNotPresent {
-		Optional<TitleEntity> entity = titleDao.findById(titleDto.getId());
+	public TitleEntity update(TitleEntity titleEntity) throws EntityNotFoundException {
+		Optional<TitleEntity> entity = titleDao.findById(titleEntity.getId());
 		if (!entity.isPresent()) {
-			throw new EntityNotPresent(titleDto.getId(), "Title does not exist!");
+			throw new EntityNotFoundException("Title does not exist!");
 		}
-		TitleEntity t = titleDao.save(titleMapper.toTitle(titleDto));
-		return titleMapper.toTitleDto(t);
-		
+		TitleEntity t = titleDao.save(titleEntity);
+		return t;
 	}
 
 	@Override
-	public Optional<TitleDto> findById(Long id) {
+	public Optional<TitleEntity> findById(Long id) {
 		Optional<TitleEntity> title = titleDao.findById(id);
 		if (title.isPresent()) {
-			return Optional.of(titleMapper.toTitleDto(title.get()));
+			return Optional.of(title.get());
 		}
 		return Optional.empty();
 	}
